Create num int
This code defines a function integral that calculates the numerical integral of a function using the method of successive rectangles (also known as the rectangle rule). Then, a simple function f is defined to implement a mathematical function. Finally, the necessary inputs for calculating the integral are taken from the user, and the result of the calculation is printed.